requirements/base.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
dj-database-url>=2.2,<3.0
pymongo>=4.6,<5.0
firebase-admin>=6.4,<7.0
python-dotenv>=1.0,<2.0
gunicorn>=21.2,<22.0
uvicorn[standard]>=0.29,<1.0
redis>=5.0,<6.0
requests>=2.32,<3.0
psycopg[binary]>=3.2,<4.0
whitenoise[brotli]>=6.9,<7.0
